Roteiro2-AllIn/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
dnspython==2.7.0
idna==3.10
pyfiglet==1.0.2
python-dateutil==2.9.0.post0
python-whois==0.9.5
requests==2.32.3
scapy==2.6.1
six==1.17.0
urllib3==2.4.0
